acme-common/requirements.txt
alabaster==0.7.12
astroid==2.9.3
attrs==21.4.0
Babel==2.9.1
bandit==1.7.1
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
coverage==6.2
decopatch==1.4.8
docutils==0.17.1
execnet==1.9.0
flake8==4.0.1
gitdb==4.0.9
GitPython==3.1.26
graphviz==0.19.1
idna==3.3
imagesize==1.3.0
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
makefun==1.13.1
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
pbr==5.8.0
pipdeptree==2.2.1
platformdirs==2.4.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.11.2
pylint==2.12.2
pyparsing==3.0.6
pytest==6.2.5
pytest-asyncio==0.17.2
pytest-cases==3.6.8
pytest-cov==3.0.0
pytest-flake8==1.0.7
pytest-forked==1.4.0
pytest-mock==3.6.1
pytest-xdist==2.5.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==4.4.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stevedore==3.5.0
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.8
wrapt==1.13.3
xlrd==2.0.1
